Don't download composer in cross-compile builds

+++ compile.sh
@@ -926,23 +926,27 @@ fi
 
 
 #Composer
-echo -n "[Composer] downloading..."
-EXPECTED_SIGNATURE=$(download_file https://composer.github.io/installer.sig)
-download_file https://getcomposer.org/installer > composer-setup.php
-ACTUAL_SIGNATURE=$($DIR/bin/php7/bin/php -r "echo hash_file('SHA384', 'composer-setup.php');")
+if [ "$IS_CROSSCOMPILE" != "yes" ]; then
+	echo -n "[Composer] downloading..."
+	EXPECTED_SIGNATURE=$(download_file https://composer.github.io/installer.sig)
+	download_file https://getcomposer.org/installer > composer-setup.php
+	ACTUAL_SIGNATURE=$($DIR/bin/php7/bin/php -r "echo hash_file('SHA384', 'composer-setup.php');")
 
-if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
-then
-    >&2 echo ' ERROR: Invalid Composer installer signature'
-    echo 'ERROR: Invalid Composer installer signature' >> "$DIR/install.log" 2>&1
-    rm composer-setup.php
-    exit 1
+	if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
+	then
+		>&2 echo ' ERROR: Invalid Composer installer signature'
+		echo 'ERROR: Invalid Composer installer signature' >> "$DIR/install.log" 2>&1
+		rm composer-setup.php
+		exit 1
+	fi
+
+	echo -n " installing..."
+	$DIR/bin/php7/bin/php composer-setup.php --install-dir=bin >> "$DIR/install.log" 2>&1
+	rm composer-setup.php
+	echo " done!"
+else
+	echo "WARNING: Can't get Composer for cross-compile builds, you will need to install Composer manually on the target machine."
 fi
-
-echo -n " installing..."
-$DIR/bin/php7/bin/php composer-setup.php --install-dir=bin >> "$DIR/install.log" 2>&1
-rm composer-setup.php
-echo " done!"
 
 
 date >> "$DIR/install.log" 2>&1
